toollib: rework process_each_lv
A new flag from commands makes explicit when
process_each_lv may query for "all vgs" and
run the command on each.  A command is run
against all vgs only when this flag is
provided and no vg/lv names or tags are provided.

When vg/lv names or tags are explicitly named
on the command line, the command is only run
against named vgs/lvs, or vgs/lvs with tags
matching one provided.

There should be no functional change.