toollib: remove replicator wrappers in vg processing
The cmd_vg wrappers were not serving any purpose
because a single vg was only ever added to the
cmd_vg set.  Remove the cmd_vg layer in
process_each_vg to simplify the code.

Pull the previous _process_one_vg into the calling
function and clarify some of the logic.