toollib: rework process_each_pv
Split up process_each_pv, matching the previous
renovation to process_each_vg,lv.

The VG_GLOBAL lock is now taken by the two callers
that had wanted it.

Two new flags passed to process_each_pv control
which pvs should be processed when command args
contain no pv names or tags:

ENABLE_ALL_VGNAMES
Scan for all vgs, including orphan vgs, and process each pv in them.

ENABLE_ALL_DEVS
Scan for all devices, including non-pvs, and process each.

process_single_pv is always passed the vg
when the pv is not an orphan.  If the pv is
an orphan (or has no mda and is not otherwise
identified as belonging to a vg), vg is NULL.

Previously, process_single_pv would not receive
the vg in what was the ENABLE_ALL_DEVS case,
and had to read the vg itself.  This case is now
removed from instances of process_single_pv.