toollib: close vg devs after process_single_vg
This test was failing at vg1:
vgsplit-stacked.sh: vgchange -a n $vg $vg1

because a pv in $vg was an lv from $vg1, and
because the devs in $vg were still open after
processing $vg, the lv from $vg1 was still open,
causing vgchange to fail on $vg1 with:

Can't deactivate volume group "@PREFIX@vg1" with 1 open logical volume(s)

I'm not sure that closing the devs this way is
correct, because I've not been able to make sense
of the way devices are closed otherwise.  There
seems to be a liberal sprinkling of dev_close_all().
I also don't see how the devs would have been
closed before the recent changes, but apparently
they were because this test was passing.