refactor: provide better return values for dev_get_primary_dev fn and enhance partition detection
dev_get_primary_dev now returns:

 0 if the dev is already a primary dev
 1 if the dev is a partition, primary dev is returned in "result" (output arg)
 -1 on error

This way, we can better differentiate between the error state
and the state in which the dev supplied is not a partition
(which was same before) in the caller.

Also, if we already have information about the device type,
we can chec its major number against the dev_type_register
directly, we don't need to go through the sysfs - we only check
the major:minor pair which is a bit more straightforward and faster.
If the dev_type_register does not have info about this device
type, the code just fallbacks to the original sysfs interface
to get the partition info.