filter-mpath: recognize partitions on mpath components
If multipath components are filtered, we need to filter out any
partitions found on these components as well to avoid any ambiguity
as mappings are also created for these partitions on top of the mpath
device itself (e.g. by means of calling kpartx for the mpath device
which should happen automatically by means of udev rules once the
mpath device is created).

For example, let's consider consider this configuration:

        (LVM)
          |
      mpath-test1
          |
      mpath-test
          |
sda1 ---------- sdb1
   \ |        |/
    sda      sdb

E.g. for sda1 and sdb1, the code will detect this and skip
the partition that is a part of the multipath component:

  <snippet from the log>
    #filters/filter-mpath.c:156         /dev/sda1: Device is a partition, using primary device /dev/sda for mpath component detection
    130 #ioctl/libdm-iface.c:1724         dm status   (253:2) OF [16384](*1)
    131 #filters/filter-mpath.c:196         /dev/sda1: Skipping mpath component device
  </snippet from the log>