toollib: rework process_each_lv
The purpose of the series reworking process_each functions
is to split process_each into "breadth first" phases rather
than "depth first".  Instead of fully processing each item
before starting to process the next, the same phase of
processing is applied to all.  This approach allows for
cleaner, simpler, and smarter processing.  It will also
allow locks to be acquired more easily.

This second patch in the series mainly splits collecting the
arguments from the processing of the arguments, like the
previous patch to process_each_vg.

A new flag from commands makes explicit when process_each_lv
may query for "all vgs" and run the command on each.
A command is run against all vgs only when this flag is
provided and no vg/lv names or tags are provided.

Different commands may or may not want "no args" to expand
to "all vgs", so it is better to make the desired behavior
explicit.  The behavior of "no args" remains the same but
is now explicit.

As before, when vg/lv names or tags are explicitly named on the
command line, the command is only run against named vgs/lvs, or
vgs/lvs with tags matching one provided.

There should be no functional change.