toollib: read vgs before processing
The purpose of the series reworking process_each functions
is to split process_each into "breadth first" phases rather
than "depth first".  Instead of fully processing each item
before starting to process the next, the same phase of
processing is applied to all.  This approach allows for
cleaner, simpler, and smarter processing.  It will also
allow locks to be acquired more easily.

This third patch in the series reads all vgs in a single
stage before processing each one.

Locking is currently hidden within vg_read,
so another effect of this change is that all
locks are acquired and held together at the
start, rather than being acquired during the
processing.

This is a step toward the introduction of a new
phase that acquires all locks explicitly before
the phase that reads each vg.