toollib: rewrite process_each_pv
The purpose of the series reworking process_each functions
is to split process_each into "breadth first" phases rather
than "depth first".  Instead of fully processing each item
before starting to process the next, the same phase of
processing is applied to all.  This approach allows for
cleaner, simpler, and smarter processing.  It will also
allow locks to be acquired more easily.

This fourth patch in the series converts process_each_pv
to the same model as the new versions of process_each_vg,lv:
read a vg, process each pv in it.

There is an extra loop at the end to process devs
that are not pvs, for commands that want to include
those.

The process_single function for a specific command should
just use the vg and pv args that are provided from the
process_each loop.