report: select: add dm_report_set_output_selection to libdm
The selection is set by using the dm_report_set_output_selection
in the form of a string that contains field names to check against and
selection operators. Then the function reads and it parses the selection
string and composes a selection tree out of it for checking against
individual fields. The parsed selection tree is stored in dm_report
structure as "selection_root".