toollib: command should fail if ignoring named arg
If a foreign VG is ignored when it's included by "all vgs",
then the command shouldn't fail.

If a foreign VG is ignored when it's named explicitly as
a command arg, then the command should fail.

Also, remove ignore_vg from reporter functions because it
repeats what has been done in process_each given the recent
new version of process_each.