lvmlockd: vgcreate/vgremove call init_vg/free_vg
vgcreate calls lockd_init_vg() to do any create/initialize
steps that are needed in lvmlockd for the given lock_type.

vgcreate calls lockd_free_vg_before() to do any removal/freeing
steps that are needed in lvmlockd for the given lock_type
before the VG is removed on disk.

vgcreate calls lockd_free_vg_final() to do any removal/freeing
steps that are needed in lvmlockd for the given lock_type
after the VG is removed on disk.

When the lock_type is sanlock, the init/free also include
lvm client side steps to create/remove an internal LV on
which sanlock will store the locks for the VG.