lvmlockd: start and stop VG lockspace
A lock manager requires an application to "start" or "join"
a lockspace before using locks from it.  Start is the point
at which the lock manager on a host begins interacting with
other hosts to coordinate access to locks in the lockspace.
Similarly, an application needs to "stop" or "leave" a
lockspace when it's done using locks from the lockspace so
the lock manager can shut down and clean up the lockspace.

lvmlockd uses a lockspace for each sanlock|dlm VG, and the
lockspace for each VG needs to be started before lvm can use
it.  These commands tell lvmlockd to start or stop the
lockspace for a VG:

vgchange --lock-start vg_name
vgchange --lock-stop vg_name

To start the lockspace for a VG, lvmlockd needs to know which
lock manager (sanlock or dlm) to use, and this is stored in the
VG metadata as lock_type = "sanlock|dlm", along with data that
is specific to the lock manager for the VG, saved as lock_args.
For sanlock, lock_args is the location of the locks on disk.
For dlm, lock_args is the name of the cluster the dlm should use.

So, the process for starting a VG includes:

- Reading the VG without a lock (no lock can be acquired
  because the lockspace is not started).

- Taking the lock_type and lock_args strings from the
  VG metadata.

- Asking lvmlockd to start the VG lockspace, providing
  the lock_type and lock_args strings which tell lvmlockd
  exactly which lock manager is needed.

- lvmlockd will ask the specific lock manager to join the
  lockspace.

The VG read in the first step, without a lock, is not used for
for anything except getting the lock information needed to start
the lockspace.  Subsequent use of the VG would use the VG lock.

In the case of a sanlock VG, there is an additional step in the
sequence.  Between the second and third steps, the vgchange
lock-start command needs to activate the internal LV in the VG
that holds the sanlock locks.  This LV must be active before
sanlock can join the lockspace.

Starting and stopping VG's would typically be done automatically
by the system, similar to the way LV's are automatically activated
by the system.  But, it is always possible to directly start/stop VG
lockspaces, as it is always possible to directly activate/deactivate
LVs.  Automatic VG start/stop will be added by a later patch, using
the basic functionality from this patch.