lvmlockd: VG locking
The VG lock is a simple read/write lock that protects a VG's
metadata.  The VG lock is used in shared mode to read the VG
and in exclusive mode to modify the VG.

The function to acquire/release the VG lock is lockd_vg():

. To acquire the vg lock in ex mode for writing:
  lockd_vg(cmd, vg_name, "ex", 0);

. To acquire the vg lock in sh mode for reading:
  lockd_vg(cmd, vg_name, "sh", 0);

. To release the vg lock:
  lockd_vg(cmd, vg_name, "un", 0);

The lockd_vg() function sends a message to lvmlockd, asking for
the lock in the specified mode.  lvmlockd acquires the lock from
the underlying lock manager, and sends the result back to the command.
When the command exits, or calls lockd_vg("un"), lvmlockd releases
the lock in the underlying lock manager.

When lvm is compiled without lvmlockd support, all the lockd_vg calls
simply compile to success (1).

Using the vg lock in commands is simple:

. A command that wants to read the VG should acquire the vg lock
  in the sh mode, then read the VG:

  lockd_vg(cmd, vg_name, "sh", 0);
  vg = vg_read(cmd, vg_name);
  lockd_vg(cmd, vg_name, "un", 0);

. A command that wants to write the VG should acquire the vg lock
  in the ex mode, then make changes to the VG and write it:

  lockd_vg(cmd, vg_name, "ex", 0);
  vg = vg_read(cmd, vg_name);
  ...
  vg_write(vg);
  vg_commit(vg);
  lockd_vg(cmd, vg_name, "un", 0);

When a command processes multiple VGs, e.g. using toollib process_each,
then VG lock should be explicitly unlocked when the command is done
processing the VG, i.e. lockd_vg(cmd, vg_name, "un", 0) as shown above.

When a command processes a single VG or pair of VGs, then the command can
simply exit and lvmlockd will automatically unlock the VG lock(s) that the
command had acquired.

Locking conflicts:

When a command calls lockd_vg(), the lock request is passed to lvmlockd.
lvmlockd makes the corresponding lock request in the lock manager using a
non-blocking request.  If another command on another host holds the vg
lock in a conflicting mode, the lock request fails, and lvmlockd returns
the failure to the command.  The command reports the lock conflict and
fails.

A future option may enable lvmlockd to automatically retry lock requests
that fail due to conflicts with locks of commands running concurrently on
other hosts.  (These retries could be disabled or limited to a certain
number via a command or config option.)  This way, simple, transient
locking conflicts between commands on different hosts would be hidden.

Caching:

lvmlockd uses the lvb in the VG lock to hold the VG seqno.  When a command
writes VG metadata with a new seqno (under an ex lock), it sends lvmlockd
the new VG seqno by calling lockd_vg_update(vg).  When lvmlockd unlocks
the ex VG lock, it saves this new seqno in the vg lock's lvb.  When other
hosts next acquire the VG lock, they will read the lvb, see the new seqno
is higher than the last seqno they saw, and know that their cached copy of
the VG is stale.  When lvmlockd sees this, it invalidates the cached copy
of the VG in lvmetad.  When a command then reads the VG from lvmetad, it
will see that it's stale, will reread the latest VG from disk, and update
the cached copy in lvmetad.

These commands do not yet work with lvmlockd lock_types (sanlock|dlm):
. vgsplit
. vgmerge
. vgrename
. lvrename