lvmlockd: LV locking
An lv lock is a simple ex/sh lock that is:
- acquired before the LV is activated
- released after the LV is deactivated

If the lock cannot be acquired, the LV is not activated.

Unlike gl and vg locks, lv locks are persistent; they
remain after a command acquiring them ends.  They are
held by the lvmlockd process.  lvmlockd uses persistent
locks in the lock manager so if lvmlockd exits, the
lv locks are not dropped, and the LVs remain protected.

The lvchange/vgchange -a option gains a new "s" arg to
represent shared locks ("e" already means exclusive):
-aey = activate using an ex lock
-asy = activate using a sh lock

An unspecified activation mode, e.g. -ay, defaults to the
ex mode in lvmlockd VGs.

If one host has an LV activated with an ex lock, any other
host that attempts to activate the LV (in any mode) will
fail to acquire the lock and will report the error.  If
a host has an LV activated with a sh lock, any other host
can also activate that LV with a sh lock, but any attempt
to activate the LV with an ex lock will fail.

Commands that modify an LV (as opposed to changing the
activation state) also use lv locks.  These are lvconvert
and other variations of lvchange.  When the LV is already
active (and locked), the lock requests do nothing, but if
the LV is not active, these commands acquire the LV lock
prior to changing the LV.  If the LV is active and locked
on another host, the lvchange/lvconvert command will fail
when the LV lock request fails.

Removing an LV with lvremove must also acquire the lv lock
prior to removing the LV.  If the LV is active and locked
on another host, lvremove will fail.

A thin pool LV has an lv lock that applies to all usage of
that pool.  Thin LVs do not have individual LV locks.

A cache pool LV does not have an independent lv lock.  When the
cache pool LV is linked to an origin LV, the lock of the origin
LV will protect the combined origin + cache pool.