vgchange: allow lock_type to be changed
When lvm is using clvm, the following are possible:
- change lock type from none to clvm
- change lock type from clvm to none

When lvm is using lvmlockd, the following are possible:
- change lock type from none to clvm (with warning)
- change lock type from clvm to none
- change lock type from none to a lockd type (sanlock|dlm)
- change lock type from clvm to a lockd type (sanlock|dlm)
- change lock type from lockd type to none (TODO)
- change lock type from lockd type to clvm (TODO)

The TODO variations are still missing the steps to
undo/reverse the existing lockd type, so are currently
disabled.

A special 'vgchange --lock-type none --force' can be used
to forcibly clear all locking settings from the VG metadata,
skipping any steps that would normally be done to cleanly
undo/reverse the existing lock type.