vgchange: lock-start options for auto and wait
Adds options to use with --lock-start,
--lock-opt wait|auto|autowait

wait: wait for the start to finish
auto: use when the system is running the command
autowait: both auto and wait

The auto option enables the use of lvm.conf activation_lock_start_list
which defines VG's that should be automatically started by the system.
This is similar to the auto_activation_volume_list used when the system
automatically activates LV's.