lvmlockd: special status for lockd VGs and LVs
Previous versions of lvm don't recognize lock_type in
lockd VGs, so they would allow a lockd VG to be used
as a local VG without restriction.  To prevent this,
the WRITE flag is removed from lockd VGs and replaced
with the WRITE_LOCKD flag.  This causes old versions
of lvm to think the VG is not writable.  The same
is done for LV flags.

WRITE_LOCKD is put on the flags line of the metadata
instead of the status line.  If it was put on the status
line, old versions of lvm would ignore the VG because of
the unrecognized flag, and the PVs in the VG would appear
unused.

In memory, lvm continues to just use the WRITE flag with
vg->status and lv->status.  WRITE is swapped with WRITE_LOCKD
when a lockd VG is read/written to/from disk.

The hidden/internal lvmlock LV is given the special LV
flag like other special lvs.  This does not go to disk.