lvmlockd: special status for lockd VGs and LVs
Previous versions of lvm don't recognize lock_type in
lockd VGs, so they would allow a lockd VG to be used
as a local VG without restriction.  So prevent this,
the WRITE flag is removed from lockd VGs and replaced
with the WRITE_LOCKD flag.  This causes old versions
of lvm to think the VG is not writable.

Similarly, old versions need to be prevented from
writing to LVs in lockd VGs (a read only VG does not
prevent lvm from writing to LVs in it).  So, LVs in
lockd VGs have WRITE replaced by WRITE_LOCKD.

In both cases, WRITE_LOCKD is put on the flags line of
the metadata instead of the status line.  If it was put
on the status line, old versions of lvm would ignore the
VG because of the unrecognized flag, and the PVs in the
VG would appear unused.

In memory, lvm continues to only use the WRITE flag with
lv->status.  WRITE is swapped with WRITE_LOCKD when the
LV is read/written on disk.

In memory, lvm uses both WRITE and WRITE_LOCKD with
vg->status as appropriate.

The hidden/internal lvmlock LV is given the special LV
flag like other special lvs.  This does not go to disk.