lvmlockd: lock_type and lvmlockd setup
The locking required to access a VG is a property of the VG,
and is specified in the VG metadata as the "lock_type".

When lvm sees a VG, it looks at the VG's lock_type to determine
if locks are needed and from where:

- If the VG has no lock_type, or lock_type "none", then no locks
  are needed.

  This a "local VG".  If the VG is visible to multiple hosts,
  the VG system_id provides basic protection.  A VG with an
  unmatching system_id is inaccessible.

- If the VG has lock_type "sanlock" or "dlm", then locks
  are needed from lvmlockd, which acquires locks from either
  sanlock or dlm respectively.

  This is a "lockd VG".  If lvmlockd or the supporting lock
  manager are not running, then the lockd VG is inaccessible.

- If the VG has the CLUSTERED status flag (or lock_type "clvm"),
  then locks are needed from clvmd.

  This is a "clvm VG".  If clvmd or the supporting clustering or
  locking are not running, then the clvm VG is inaccessible.

Settings in lvm.conf tell lvm commands which locking daemon to use:

- global/use_lvmlockd=1: tells lvm to use lvmlockd when accessing
  VGs with lock_type sanlock|dlm.

- global/locking_type=3: tells lvm to use clvmd when accessing
  VGs with CLUSTERED flag (or lock_type clvm).

LVM commands cannot use both lvmlockd and clvmd at the same time:

- use_lvmlockd=1 should be combined with locking_type=1
- locking_type=3 (clvmd) should be combined with use_lvmlockd=0

So, different configurations allow access to different VG's:

- When configured to use lvmlockd, lvm commands can access VG's
  with lock_type sanlock|dlm, and VG's with CLUSTERED are ignored.

- When configured to use clvmd (locking_type 3), lvm commands
  can access VG's with the CLUSTERED flag, and VG's with
  lock_type sanlock|dlm are ignored.

- When configured to use neither lvmlockd nor clvmd, lvm commands
  can access only local VG's.  lvm will ignore VG's with lock_type
  sanlock|dlm, and will ignore VG's with CLUSTERED (or lock_type clvm).

A VG is created with a specific lock_type:

- vgcreate --lock_type <arg> is a new syntax that can specify the
  lock_type directly.  <arg> may be: none, clvm, sanlock, dlm.

  sanlock|dlm require lvmlockd to be configured (in lvm.conf) and running.
  clvm requires clvmd to be configured (in lvm.conf) and running.

- vgcreate --clustered y (or -cy) is the old syntax that still works,
  but it is not preferred because the lock_type is not explicit.

  When clvmd is configured, -cy creates a VG with lock_type clvm.
  When lvmlockd is configured, -cy creates a VG with lock_type sanlock,
  but this can be changed to dlm with lvm.conf vgcreate_cy_lock_type.

Notes:

In the VG metadata, the lock_type string is accompanied by
a lock_args string.  The lock_args string is lock-manager-specific
data associated with the VG.  For sanlock, the location on disk
of the locks, or for dlm, the cluster name.

In a VG with lock_type sanlock|dlm, each LV also has a lock_type
and lock_args in the metadata.  The LV lock_type currently always
matches the lock_type of the VG.  For sanlock, the LV lock_args
specify the disk location of the LV lock.